refactor: api 로직 엔티티로 이동, query 훅 model 하위로 이동
